Develop

+++ Game.Application/Services/GameService.cs
@@ -1,32 +1,41 @@
-﻿using AutoMapper;
-using Game.Domain.Interfaces.Repositories;
+﻿using Game.Domain.Interfaces.Repositories;
 using Game.Domain.Interfaces.Services;
 using Game.Domain.Model;
+using Game.Domain.Resources;
 using Game.Infra.Data.Core;
+using System.Net;
 
 namespace Game.Application.Services
 {
     internal class GameService : IGameService
     {
-        private readonly IMapper _mapper;
         private readonly IGameRepository _gameRepository;
 
-        public GameService(IMapper mapper, IGameRepository gameRepository)
+        public GameService(IGameRepository gameRepository)
         {
-            _mapper = mapper;
             _gameRepository = gameRepository;
         }
 
         public CustomResult GetById(int id)
         {
+            if (id < 1)
+                return CustomResult.Fail($"{Messages.INVALID_BOARD_ID} {id}");
+
             var result = _gameRepository.GetById(id);
+            if (result == null)
+                return CustomResult.Fail($"{Messages.BOARD_STATE_NOT_FOUND_WITH_ID} {id}", HttpStatusCode.NotFound);
 
             return CustomResult.Ok(result);
         }
 
         public CustomResult GetFinalState(int iterations)
         {
+            if (iterations < 1)
+                return CustomResult.Fail(Messages.INVALID_ITERATIONS);
+
             var result = _gameRepository.GetFinalState(iterations);
+            if (result == null)
+                return CustomResult.Fail(Messages.INVALID_FINAL_STATE);
 
             return CustomResult.Ok(result);
         }
@@ -34,18 +43,33 @@ namespace Game.Application.Services
         public CustomResult GetNextState()
         {
             var result = _gameRepository.GetNextState();
+            if (result == null)
+                return CustomResult.Fail(Messages.INVALID_NEXT_STATE);
+
             return CustomResult.Ok(result);
         }
 
         public CustomResult RemoveById(int id)
         {
+            if (id < 1)
+                return CustomResult.Fail($"{Messages.INVALID_BOARD_ID} {id}");
+
             var result = _gameRepository.RemoveById(id);
+            if (result == null)
+                return CustomResult.Fail($"{Messages.BOARD_STATE_NOT_FOUND_WITH_ID} {id}", HttpStatusCode.NotFound);
+
             return CustomResult.Ok(result);
         }
 
         public CustomResult Simulate(int iterations)
         {
+            if (iterations < 1)
+                return CustomResult.Fail(Messages.INVALID_ITERATIONS);
+
             var result = _gameRepository.Simulate(iterations);
+            if (result == null)
+                return CustomResult.Fail($"{Messages.CURRRENT_BOARD_STATE_NOT_FOUND}", HttpStatusCode.NotFound);
+
             return CustomResult.Ok(result);
         }
